Extract an AnimationList UI component.
